Order

+++ APIs/PTP.Application/Features/Menus/Commands/CreateMenuCommand.cs
@@ -50,14 +50,35 @@ public class CreateMenuCommand:IRequest<MenuViewModel>
         public async Task<MenuViewModel> Handle(CreateMenuCommand request, CancellationToken cancellationToken)
         {
             _logger.LogInformation("Create Menu:\n");
+            TimeSpan.TryParseExact(request.CreateModel.StartTime, @"hh\:mm", CultureInfo.InvariantCulture, out TimeSpan startTime);
+            TimeSpan.TryParseExact(request.CreateModel.EndTime, @"hh\:mm", CultureInfo.InvariantCulture, out TimeSpan endTime);
+            if (startTime>=endTime) throw new BadRequestException("Start Time must higher than End Time");
+
+            if(!await CheckTime(request.CreateModel.StoreId,startTime,endTime)) throw new BadRequestException("Time is not valid!");
             var menu= _mapper.Map<Menu>(request.CreateModel);
-            menu.StartTime=TimeSpan.ParseExact(request.CreateModel.StartTime, @"hh\:mm", CultureInfo.InvariantCulture);
-            menu.EndTime=TimeSpan.ParseExact(request.CreateModel.EndTime, @"hh\:mm", CultureInfo.InvariantCulture);
+            
 
             await _unitOfWork.MenuRepository.AddAsync(menu);
             if( !await _unitOfWork.SaveChangesAsync()) throw new BadRequestException("Save changes Fail!");
                 await _cacheService.RemoveByPrefixAsync(CacheKey.MENU);
             return _mapper.Map<MenuViewModel>(menu);
+        }
+
+        private async Task<bool> CheckTime(Guid storeId,TimeSpan sTime, TimeSpan eTime) 
+        {
+            var menus = await _unitOfWork.MenuRepository.WhereAsync(x =>
+                                 x.StoreId == storeId);
+            
+            if(menus.Count==0) return true;
+
+            foreach (var item in menus)
+            {
+                TimeSpan.TryParseExact(item.StartTime, @"hh\:mm", CultureInfo.InvariantCulture, out TimeSpan startTime);
+                TimeSpan.TryParseExact(item.EndTime, @"hh\:mm", CultureInfo.InvariantCulture, out TimeSpan endTime);
+                if (startTime <= sTime && endTime > sTime) throw new BadRequestException($"Start time is duplicate with menu-{item.Id}" );
+                if (startTime < eTime && endTime >= eTime) throw new BadRequestException($"End time is duplicate with menu-{item.Id}" );
+            }
+            return true;
         }
     }
 }
